+++ contracts/src/multisig/examples/test/simulators/ForwarderUnshieldedSimulator.ts
@@ -5,11 +5,14 @@ import {
 import {
   type ContractAddress,
   type Either,
-  Contract as ForwarderUnshielded,
+  Contract as ForwarderUnshieldedExample,
   ledger,
   type UserAddress,
-} from '../../../../../artifacts/ForwarderUnshielded/contract/index.js';
-import { EmptyPrivateState, emptyWitnesses } from '../../EmptyWitnesses.js';
+} from '../../../../../artifacts/ForwarderUnshieldedExample/contract/index.js';
+import {
+  EmptyPrivateState,
+  emptyWitnesses,
+} from '../../../test/EmptyWitnesses.js';
 
 type ForwarderUnshieldedArgs = readonly [parent: UserAddress];
 
@@ -17,16 +20,16 @@ const ForwarderUnshieldedSimulatorBase = createSimulator<
   EmptyPrivateState,
   ReturnType<typeof ledger>,
   ReturnType<typeof emptyWitnesses>,
-  ForwarderUnshielded<EmptyPrivateState>,
+  ForwarderUnshieldedExample<EmptyPrivateState>,
   ForwarderUnshieldedArgs
 >({
   contractFactory: (witnesses) =>
-    new ForwarderUnshielded<EmptyPrivateState>(witnesses),
+    new ForwarderUnshieldedExample<EmptyPrivateState>(witnesses),
   defaultPrivateState: () => EmptyPrivateState,
   contractArgs: (parent) => [parent],
   ledgerExtractor: (state) => ledger(state),
   witnessesFactory: () => emptyWitnesses(),
-  artifactName: 'ForwarderUnshielded',
+  artifactName: 'ForwarderUnshieldedExample',
 });
 
 export class ForwarderUnshieldedSimulator extends ForwarderUnshieldedSimulatorBase {